+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Removes backward compatibility for MAPL_FlapCLI functions. Only accepts function usage in which the result is of
+- Removes backward compatibility for MAPL_FlapCLI and MAPL_FargparseCLI functions. Only accepts function usage in which the result is of
   MAPL_CapOptions type.
 
 ### Added